split watchmedo logic into _watchmedo.py (yaml and argh independent)
